Created processor objects for ref data and stock
