+++ spark-scala-demo/src/main/scala/joe/spark/scala/package.scala
@@ -1,0 +1,24 @@
+package joe.spark
+
+package object scala {
+  
+  final case class StockRecord(exchange: String, 
+                      stockSymbol: String, 
+                      date: String, 
+                      stockPriceOpen: Double, 
+                      stockPriceHigh: Double, 
+                      stockPriceLow: Double, 
+                      stockPriceClose: Double, 
+                      stockVolume: Int, 
+                      stockPriceAdjClose: Double)
+                      
+  final case class State(id: Long, 
+                      name: String, 
+                      code: String, 
+                      countryId: Long)                      
+
+  final case class Country(id: Long, 
+                      name: String, 
+                      code: String)                      
+  
+}
